Move utils to a separate file
